tools: Extract png generation code to separate module.
